+++ test/e2e/Apps/BasicPython/function_app.py
@@ -20,6 +20,7 @@ from timeout_orchestration import bp as timeout_orchestration_bp
 from purge_orchestration_history import bp as purge_orchestration_history_bp
 from class_based_entities import bp as class_based_entities_bp
 from is_replaying_checks import bp as is_replaying_checks_bp
+from disabled_functions import bp as disabled_functions_bp
 
 app = func.FunctionApp(http_auth_level=func.AuthLevel.ANONYMOUS)
 
@@ -59,3 +60,4 @@ app.register_blueprint(timeout_orchestration_bp)
 app.register_blueprint(purge_orchestration_history_bp)
 app.register_blueprint(class_based_entities_bp)
 app.register_blueprint(is_replaying_checks_bp)
+app.register_blueprint(disabled_functions_bp)